Split ucrx template implementations
