+++ src/compiler/impl/opaque.ucrx-template.ts
@@ -1,21 +1,17 @@
 import { CHURI_MODULE } from '../../impl/module-names.js';
+import { ExternalUcrxTemplate } from '../rx/external.ucrx-template.js';
 import { UcrxLib } from '../rx/ucrx-lib.js';
 import { UcrxTemplate } from '../rx/ucrx-template.js';
-import { VoidUcSchema } from './void.uc-schema.js';
 
-export class OpaqueUcrxTemplate extends UcrxTemplate {
+export class OpaqueUcrxTemplate extends ExternalUcrxTemplate {
 
   constructor(lib: UcrxLib) {
     super({
       lib,
-      schema: VoidUcSchema,
+      importFrom: CHURI_MODULE,
       className: 'OpaqueUcrx',
       args: [],
     });
-  }
-
-  override get base(): string {
-    return CHURI_MODULE;
   }
 
   protected override declareMethods(): UcrxTemplate.MethodDecls {
