--- src/compiler/impl/void.uc-schema.ts
@@ -1,7 +1,0 @@
-import { asis } from '@proc7ts/primitives';
-import { UcSchema } from '../../schema/uc-schema.js';
-
-export const VoidUcSchema: UcSchema<void> = {
-  type: 'void',
-  asis,
-};
