+++ src/schema/uc-list.spec.ts
@@ -259,14 +259,13 @@ describe('UcList', () => {
       try {
         await lib.compile().toDeserializers();
       } catch (e) {
+        console.error(e);
         error = e as UnsupportedUcSchemaError;
       }
 
       expect(error).toBeInstanceOf(UnsupportedUcSchemaError);
       expect(error?.schema.type).toBe('test-type');
-      expect(error?.message).toBe(
-        'readList$deserialize: Can not deserialize list item of type "test-type"',
-      );
+      expect(error?.message).toBe('test-type[]: Can not deserialize list item of type "test-type"');
       expect(error?.cause).toBeInstanceOf(UnsupportedUcSchemaError);
       expect((error?.cause as UnsupportedUcSchemaError).schema.type).toBe('test-type');
     });
